Expose formula copy visitor
Copy visitor will be used by consumers to do things like shift
references and so on. That is not currently possible, because
RefModVisitor doesn't expose methods to modify references and even if
it did, there is idea to know if reference is from sheet reference or
3D reference.

Also, add info about current sheet to the context, so visitor knows
implicit sheet for non-sheet references (e.g. pure `A1`). In some
places, I have used string.Empty, because that will never conflict
with any sheet name (empty name is invalid sheet name).