LOG10 function is not a cell func
Parser accepts cell references as functions, e.g. $A$1(45). When formula
is converted between A1 and R1C1, make sure the LOG10 is *not*
interpreted as a cell function. That would  cause all sorts of problems
with shifting or deleting areas.

Example: LOG10 would be turned to R1C1, the reference would shift
and then be turned back.