Add a hand-made lexer for 3.0
This is a lexer for versin 3.0 that should use Pratt parsing. The lexer
is handmade, though representable in DFA if necessary.

After experience with 2.0, it seems that too much was done in lexer
and the lexer produce far less tokens and should be far more
"relaxed". That means parser will have to deal with more relaxed tokens
that might not be valid in used context.

Performance is acceptable (150ms for tokenization of 22MB of enron
formulas), but that is because it's rather simple. A lot of logic will
be in parser.