Add an intermediate node structure to parsing
IAstFactory needs to know the whole extend of a node, e.g. to replace
it.

Example: factory gets BinaryNode callback and it should replace a node
with a result. The binary node might hav ebeen deeply nested. The range
parameter of callbacks is used to know the whole extent of original formula
that was used to create the binary node. Thus callback can combine left
side, replacement for the binary node, and left side.